Updated storage management to work in every device
