+++ src/org/digitalcampus/oppia/utils/storage/ExternalStorageStrategy.java
@@ -27,6 +27,8 @@ import android.util.Log;
 import org.digitalcampus.oppia.activity.PrefsActivity;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
 
 
 public class ExternalStorageStrategy implements StorageAccessStrategy{
@@ -36,10 +38,22 @@ public class ExternalStorageStrategy implements StorageAccessStrategy{
 
     //@Override
     public void updateStorageLocation(Context ctx){
+
+        String location = null;
         //If no mount argument passed, we set the default external mount
-        File[] dirs = ContextCompat.getExternalFilesDirs(ctx, null);
-        if (dirs.length > 0){
-            String location = dirs[dirs.length-1].toString();
+        DeviceFile external = StorageUtils.getExternalMemoryDrive();
+        if (external != null && external.canWrite()){
+            location = external.getPath();
+        }
+        else{
+            DeviceFile internal = StorageUtils.getInternalMemoryDrive();
+            if (internal != null && internal.canWrite()){
+                location = internal.getPath();
+
+            }
+        }
+        if (location != null){
+            location += getInternalBasePath(ctx);
             updateLocationPreference(ctx, location);
         }
     }
